Split out store config from Store component.
